app/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pillow
python-multipart
